perf(plugins): avoid metadata generation imports for path resolution
